Add support for parallel analysis of repositories
When running on multiple large repositories, gitlogg used to
analyze these in a sequential manner. This is fine as long as disk
speed is the bottle neck. Calling git log is a CPU expensive
operation, however, so even single SSD machines can benefit from a
performance boost by parallelizing the work.

Per default, we start n-1 sub processes, where n is the computer's
number of CPUs. This ensures we do not overload the system in case
of a desktop computer. With the -n parameter, the user can however
specify an arbitrary number of processes to spawn.

This mandated the following changes:

 - Refactoring out the core part of the script to be able to spawn
   it via xargs.

 - Refactoring out the terminal color setup to use it from both the
   main and the worker script.

 - Adding a CLI option parser and usage info.